Set up initial framework
In the beginning I wanted to set up the terminal nicely as well with
setup, but I did not find a nice way to do it that did not cause some
annoyance. Eg. creating new windows is annoying if we keep the ones from
yesterday, and it was hard to start programs in them for some cases
directly. I assume I will have a minute to set up the windows if needed
before the problem is released.