Finish day 3
Quite happy with this day. I was a bit more sleepy and off balance than
the other days, but I ended up getting a higher rank than the other
days which is nice. I was also faster than yesterday which is nice.
However I am still surprisingly slow, but I can still blame that on not
knowing Julia, such as having to look up integer division during the day
:D