Finish day 12
Not record time, at all. but still rather quick. I had some hickups with
julia and how to do stuff (for a while I thought we should do Dijkstras,
so I downloaded a PriorityQueue), but then I realized it was a simple
bfs and it went rather ok. Code is very messy but it was rather fast to
write and I did not get any real bugs (apart from not realising you
cannot go diagonaly).